requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jsonschema==4.23.0
numpy==1.26.4
alembic==1.13.2
sqlalchemy==2.0.31
requests==2.32.3
# Optional vector engines
pinecone-client==4.1.0 ; platform_system != "Windows"
weaviate-client==4.6.5
qdrant-client==1.9.2
# Optional local embeddings
sentence-transformers==3.2.1
